Data.Stream.Supply

Description

This library can be used to generate values (for example, new names) without the need to thread state. This means that functions that need to generate new values only need a supply object as an argument, and they do not need to return a new supply object as a result. This decreases the number of data-dependencies in a program, which makes it easier to exploit parallelism.

The technique for generating new values is based on the paper ''On Generating Unique Names'' by Lennart Augustsson, Mikael Rittri, and Dan Synek.

Documentation

data Supply a

newSupply :: (a -> a) -> a -> IO (Supply a)

newEnumSupply :: Enum a => IO (Supply a)

newNumSupply :: Num a => IO (Supply a)

newDupableSupply :: (a -> a) -> a -> IO (Supply a)

newDupableEnumSupply :: Enum a => IO (Supply a)

newDupableNumSupply :: Num a => IO (Supply a)

leftSupply :: Supply a -> Supply a

rightSupply :: Supply a -> Supply a

split :: Supply a -> Stream (Supply a)

splits :: Integral b => Supply a -> b -> Supply a

splitSkew :: Supply a -> Stream (Supply a)

split2 :: Supply a -> (Supply a, Supply a)

split3 :: Supply a -> (Supply a, Supply a, Supply a)

split4 :: Supply a -> (Supply a, Supply a, Supply a, Supply a)